Split renderer/settings-modal.tsx — pull shared fields and parse into parts/
